feat(blobstore): add CachingStore for block-level caching of blobs
- Implemented CachingStore that wraps a BlobStore and adds block-level caching.
- Introduced CachingBlob to handle read operations with caching.
- Added cache invalidation on blob deletion.
- Created unit tests for CachingStore and CachingBlob to ensure correct functionality.

feat(s3): implement S3 BlobStore for AWS S3 integration

- Added Store struct to implement blobstore.BlobStore for S3.
- Implemented methods for Open, Create, Delete, and List operations.
- Added s3Blob and s3WritableBlob types for handling blob operations in S3.

feat(cache): enhance LRUBlockCache with hit/miss tracking and invalidation

- Added hit and miss counters to LRUBlockCache for performance monitoring.
- Implemented Invalidate method to remove cache entries based on a predicate.
- Introduced Stats method to retrieve cache statistics.

fix(cache): update CacheKey structure to include Path for blob caching

- Modified CacheKey to include an optional Path field for better cache management.

fix(engine): integrate block caching into the Engine

- Added block cache size configuration to the Engine.
- Wrapped the existing BlobStore with CachingStore to enable caching.

fix(bm25): change count type from int to uint32 for posting struct

- Updated posting struct to use uint32 for count to prevent overflow issues.
- Adjusted related methods and tests to accommodate the new type.

test(bm25): add unit tests for MemoryIndex functionality

- Implemented tests for basic operations, deletion, and type overflow scenarios in MemoryIndex.

+++ lexical/bm25/bm25.go
@@ -17,7 +17,7 @@ const (
 
 type posting struct {
 	docID uint32
-	count int
+	count uint32
 }
 
 // MemoryIndex is a simple in-memory BM25 index.
@@ -127,7 +127,7 @@ func (idx *MemoryIndex) Add(pk model.PK, text string) error {
 	idx.docCount++
 
 	for t, count := range tf {
-		idx.inverted[t] = append(idx.inverted[t], posting{docID: docID, count: count})
+		idx.inverted[t] = append(idx.inverted[t], posting{docID: docID, count: uint32(count)})
 	}
 
 	return nil
